Aplicación Web

+++ UserManagerApi/Models/User.cs
@@ -22,13 +22,13 @@ namespace UserManagerApi.Models
         public string Address { get; set; }
 
         [Column(TypeName = "numeric(18,0)")]
-        public Decimal Phone { get; set; }
+        public long Phone { get; set; }
 
         [Column(TypeName = "nvarchar(100)")]
         public string Email { get; set; }
 
         [Column(TypeName = "numeric(3,0)")]
-        public Decimal Age { get; set; }
+        public long Age { get; set; }
 
         [Required]
         public int IdRole { get; set; }
